refactor: Service Interface 분리
